% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data}
\alias{data}
\title{Download and Load a Dataset from GitHub with Custom Name}
\usage{
data(url, new_name)
}
\arguments{
\item{url}{URL of the .rda file on GitHub}

\item{new_name}{The name to assign to the dataset in the global environment}
}
\value{
Loads the dataset into the global environment with a custom name
}
\description{
Download and Load a Dataset from GitHub with Custom Name
}
