% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_ps1_conjoint_data}
\alias{download_ps1_conjoint_data}
\title{Download and Load Conjoint Data}
\usage{
download_ps1_conjoint_data(new_name = "conjoint")
}
\arguments{
\item{new_name}{Name to assign to the loaded dataset}
}
\value{
Loads the conjoint dataset into the global environment
}
\description{
Download and Load Conjoint Data
}
