% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_ps1_logistic_data}
\alias{download_ps1_logistic_data}
\title{Download and Load ps1_logistic Data from GitHub}
\usage{
download_ps1_logistic_data()
}
\value{
Loads the ps1_logistic dataset into the global environment.
}
\description{
This function downloads the ps1_logistic.rda file from GitHub and loads it into the global environment.
}
