% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_ps1_ols_data}
\alias{download_ps1_ols_data}
\title{Download and Load ps1_ols Data from GitHub}
\usage{
download_ps1_ols_data()
}
\value{
Loads the ps1_ols dataset into the global environment.
}
\description{
This function downloads the ps1_ols.rda file from GitHub and loads it into the global environment.
}
