% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_rfm_data}
\alias{download_rfm_data}
\title{Download and Load RFM Data}
\usage{
download_rfm_data(new_name = "rfm")
}
\arguments{
\item{new_name}{Name to assign to the loaded dataset}
}
\value{
Loads the RFM dataset into the global environment
}
\description{
Download and Load RFM Data
}
