% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_rta_data}
\alias{download_rta_data}
\title{Download and Load RTA Data}
\usage{
download_rta_data(new_name = "rta")
}
\arguments{
\item{new_name}{Name to assign to the loaded dataset}
}
\value{
Loads the RTA dataset into the global environment
}
\description{
Download and Load RTA Data
}
