% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_ps1_conjoint_data}
\alias{download_ps1_conjoint_data}
\title{Download and Load conjoint data from GitHub}
\usage{
download_ps1_conjoint_data()

download_ps1_conjoint_data()
}
\value{
Loads the conjoint dataset into the global environment.

Loads the conjoint data into the global environment.
}
\description{
This function downloads the conjoint file from GitHub and loads it into the global environment.

This function downloads the conjoint file from GitHub and loads it into the global environment.
}
