% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_ps1_logistic_data}
\alias{download_ps1_logistic_data}
\title{Download and Load Cancellations Data for Logistic Regression}
\usage{
download_ps1_logistic_data(new_name = "ps1_logistic")
}
\arguments{
\item{new_name}{Name to assign to the loaded dataset}
}
\value{
Loads the cancellations dataset into the global environment
}
\description{
Download and Load Cancellations Data for Logistic Regression
}
