% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_ps1_ols_data}
\alias{download_ps1_ols_data}
\title{Download and Load Customers Data for OLS}
\usage{
download_ps1_ols_data(new_name = "ps1_ols")
}
\arguments{
\item{new_name}{Name to assign to the loaded dataset}
}
\value{
Loads the customers dataset into the global environment
}
\description{
Download and Load Customers Data for OLS
}
