% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_rfm_data}
\alias{download_rfm_data}
\title{Download and Load RFM Data from GitHub}
\usage{
download_rfm_data()
}
\value{
Loads the rfm dataset into the global environment.
}
\description{
This function downloads the rfm.rda file from GitHub and loads it into the global environment.
}
