% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_rta_data}
\alias{download_rta_data}
\title{Download and Load RFM Data from GitHub}
\usage{
download_rta_data()
}
\value{
Loads the rta dataset into the global environment.
}
\description{
This function downloads the rta.rda file from GitHub and loads it into the global environment.
}
